// frontend/src/app/model/user/user.ts
export class User {
    firstName: string | undefined;
    name: string | undefined;
    email: string | undefined;
    password: string | undefined;
    token?: string | undefined;
}
// "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhemVAYXplIiwiZXhwIjoxNjM4NzgwMzY1fQ.lXcmK_bxC8msamo-8WUvOaPsKmuJvEagbtd2enBtoQgSBNlNwnUNhcidR9HjFPyGigU1Hsb2-i-KGsh0Lobd5Q"